Unit Editor Started
Began implementing the Unit system as well as a unit editor
(tech-tree/upgrades as part of it). A large portion of the code will
need to be scrapped, as I found out you can't serialize static data in
Unity, so the method of storing all defined unit types will not work.