Unit Upgrades
Unit upgrading is implemented, although it could be better fleshed out.
Each prefab unit will be able to upgrade to a preset unit type set in
the editor.

Actually handling the upgrading is only minimally implemented; a dummy
test script with a button. Still looking into best solutions for a
reusable system.